ref distance matrix

+++ SAT.MODELS/Entities/Geolocalizacao.cs
@@ -11,7 +11,8 @@
         public string Endereco { get; set; }
         public string Pais { get; set; }
         public string Numero { get; set; }
-        public int? Duracao { get; set; }
+        // Em minutos
+        public double? Duracao { get; set; }
         public double? Distancia { get; set; }
     }
 }
